put compass object in dedicated module
